+++ src/Domain/Entities/PlannerTask.cs
@@ -41,7 +41,7 @@ public class PlannerTask : BaseAuditableEntity
     /// <summary>
     /// Gets or sets the ID of the calendar to which this task belongs.
     /// </summary>
-    public int CalendarId { get; set; }
+    public long CalendarId { get; set; }
 
     /// <summary>
     /// Gets or sets the calendar to which this task belongs.
